// frontend/src/apisv2beta1/run/models/PipelineTaskTaskPod.ts
/* tslint:disable */
/* eslint-disable */
/**
 * backend/api/v2beta1/run.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PipelineTaskTaskPodType } from './PipelineTaskTaskPodType';
import {
  PipelineTaskTaskPodTypeFromJSON,
  PipelineTaskTaskPodTypeFromJSONTyped,
  PipelineTaskTaskPodTypeToJSON,
  PipelineTaskTaskPodTypeToJSONTyped,
} from './PipelineTaskTaskPodType';

/**
 *
 * @export
 * @interface PipelineTaskTaskPod
 */
export interface PipelineTaskTaskPod {
  /**
   *
   * @type {string}
   * @memberof PipelineTaskTaskPod
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof PipelineTaskTaskPod
   */
  uid?: string;
  /**
   *
   * @type {PipelineTaskTaskPodType}
   * @memberof PipelineTaskTaskPod
   */
  type?: PipelineTaskTaskPodType;
}

/**
 * Check if a given object implements the PipelineTaskTaskPod interface.
 */
export function instanceOfPipelineTaskTaskPod(value: object): value is PipelineTaskTaskPod {
  return true;
}

export function PipelineTaskTaskPodFromJSON(json: any): PipelineTaskTaskPod {
  return PipelineTaskTaskPodFromJSONTyped(json, false);
}

export function PipelineTaskTaskPodFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): PipelineTaskTaskPod {
  if (json == null) {
    return json;
  }
  return {
    name: json['name'] == null ? undefined : json['name'],
    uid: json['uid'] == null ? undefined : json['uid'],
    type: json['type'] == null ? undefined : PipelineTaskTaskPodTypeFromJSON(json['type']),
  };
}

export function PipelineTaskTaskPodToJSON(json: any): PipelineTaskTaskPod {
  return PipelineTaskTaskPodToJSONTyped(json, false);
}

export function PipelineTaskTaskPodToJSONTyped(
  value?: PipelineTaskTaskPod | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    name: value['name'],
    uid: value['uid'],
    type: PipelineTaskTaskPodTypeToJSON(value['type']),
  };
}