// frontend/src/apisv2beta1/run/models/PipelineTaskTaskPodType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * backend/api/v2beta1/run.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 */
export const PipelineTaskTaskPodType = {
  UNSPECIFIED: 'UNSPECIFIED',
  DRIVER: 'DRIVER',
  EXECUTOR: 'EXECUTOR',
} as const;
export type PipelineTaskTaskPodType =
  (typeof PipelineTaskTaskPodType)[keyof typeof PipelineTaskTaskPodType];

export function instanceOfPipelineTaskTaskPodType(value: any): boolean {
  for (const key in PipelineTaskTaskPodType) {
    if (Object.prototype.hasOwnProperty.call(PipelineTaskTaskPodType, key)) {
      if (PipelineTaskTaskPodType[key as keyof typeof PipelineTaskTaskPodType] === value) {
        return true;
      }
    }
  }
  return false;
}

export function PipelineTaskTaskPodTypeFromJSON(json: any): PipelineTaskTaskPodType {
  return PipelineTaskTaskPodTypeFromJSONTyped(json, false);
}

export function PipelineTaskTaskPodTypeFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): PipelineTaskTaskPodType {
  return json as PipelineTaskTaskPodType;
}

export function PipelineTaskTaskPodTypeToJSON(value?: PipelineTaskTaskPodType | null): any {
  return value as any;
}

export function PipelineTaskTaskPodTypeToJSONTyped(
  value: any,
  ignoreDiscriminator: boolean,
): PipelineTaskTaskPodType {
  return value as PipelineTaskTaskPodType;
}